requirements.txt
discord.py>=1.7.3
google_api_python_client>=2.45.0
google_auth_oauthlib>=0.5.1
gspread>=5.3.2
gspread_dataframe>=3.3.0
pandas>=1.4.2
protobuf>=3.20.0
PyNaCl>=1.5.0
